Basic Render system implemented. Everything that will be rendered implements the interface Render and will be given to the renderStack in GameView. All rendered objects have an update(), getBitmap(), getX() and getY()-method. The update()-method adjusts how (bitmap) and where (X and Y) the rendered object is shown on-screen. Also added rendered Objects for Player (Figure) and Place (City). Node is next, but will probably just be a shape/line connecting to Cities. Then the players just have to move between two Cities when their currentPlace changes (instead of teleporting which happens in the mini-demo now). This can either be realised by tween or (how I will probably do it) a simple function creating a straight line between the points and moving the Figure along this function while increasing their x-value.
In the demo there are two Places and the Player has two Tickets. A button can be clicked to trigger the doTurn()-method of the Player, spending a Ticket and changing his currPlace and by that changing the x-value in the update()-Method of the Figure. He can go back once, but is out of Tickets then.